+++ takeout_mobile/lib/dialog.dart
@@ -16,22 +16,81 @@
 // along with TakeoutFM.  If not, see <https://www.gnu.org/licenses/>.
 
 import 'package:flutter/material.dart';
+import 'package:takeout_mobile/app/context.dart';
 
-void showErrorDialog(BuildContext context, String message) {
-  showDialog<void>(
-    context: context,
-    builder: (BuildContext ctx) {
-      return AlertDialog(
-        title: Text(MaterialLocalizations.of(context).alertDialogLabel),
-        content: Text(message),
-        actions: [
-          TextButton(
-            child: Text(MaterialLocalizations.of(context).okButtonLabel),
-            onPressed: () => Navigator.pop(ctx),
-          )
-        ],
-      );
-    },
-  );
+Future<void> showAlertDialog(BuildContext context, String message,
+        {void Function()? onConfirmed}) =>
+    showDialog<void>(
+      context: context,
+      builder: (BuildContext ctx) {
+        return AlertDialog(
+          title: Text(MaterialLocalizations.of(context).alertDialogLabel),
+          content: Text(message),
+          actions: [
+            TextButton(
+              child: Text(MaterialLocalizations.of(context).okButtonLabel),
+              onPressed: () {
+                Navigator.pop(ctx);
+                onConfirmed?.call();
+              },
+            )
+          ],
+        );
+      },
+    );
+
+Future<String?> textDialog(BuildContext context, String label,
+    {String? hint}) async {
+  return showDialog<String>(
+      context: context,
+      builder: (BuildContext ctx) {
+        final text = TextField(
+            onSubmitted: (value) {
+              Navigator.pop(ctx, value);
+            },
+            autofocus: true,
+            decoration: InputDecoration(labelText: label, hintText: hint));
+        return AlertDialog(
+          contentPadding: const EdgeInsets.all(16.0),
+          content: Row(
+            children: <Widget>[
+              Expanded(
+                child: text,
+              ),
+            ],
+          ),
+          actions: <Widget>[
+            TextButton(
+                child: Text(MaterialLocalizations.of(ctx).cancelButtonLabel),
+                onPressed: () {
+                  Navigator.pop(ctx);
+                }),
+          ],
+        );
+      });
 }
 
+void confirmDeleteDialog(
+    BuildContext context, String message, void Function() onConfirm) {
+  showDialog<void>(
+      context: context,
+      builder: (ctx) {
+        return AlertDialog(
+          title: Text(context.strings.confirmDelete),
+          content: Text(message),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(ctx),
+              child: Text(MaterialLocalizations.of(context).cancelButtonLabel),
+            ),
+            TextButton(
+              onPressed: () {
+                onConfirm();
+                Navigator.pop(ctx);
+              },
+              child: Text(MaterialLocalizations.of(context).okButtonLabel),
+            ),
+          ],
+        );
+      });
+}