+++ takeout_mobile/lib/home.dart
@@ -25,6 +25,7 @@ import 'package:takeout_mobile/app/app.dart';
 import 'package:takeout_mobile/app/context.dart';
 import 'package:takeout_mobile/downloads.dart';
 import 'package:takeout_mobile/link.dart';
+import 'package:takeout_mobile/playlists.dart';
 import 'package:takeout_mobile/settings/widget.dart';
 import 'package:takeout_lib/api/model.dart' hide Offset;
 import 'package:takeout_lib/art/artwork.dart';
@@ -195,6 +196,7 @@ class HomeWidget extends StatelessWidget {
         popupMenu(context, [
           PopupItem.playlist(context, (context) => _onRecentTracks(context)),
           PopupItem.popular(context, (context) => _onPopularTracks(context)),
+          PopupItem.playlists(context, (context) => _onPlaylists(context)),
           PopupItem.divider(),
           PopupItem.settings(context, (context) => _onSettings(context)),
           PopupItem.downloads(context, (context) => _onDownloads(context)),
@@ -241,6 +243,7 @@ class HomeWidget extends StatelessWidget {
 
   void _onRecentTracks(BuildContext context) {
     pushSpiff(
+        ref: '/api/activity/tracks/recent/playlist',
         context,
         (ClientCubit client, {Duration? ttl}) =>
             client.recentTracks(ttl: Duration.zero));
@@ -248,9 +251,14 @@ class HomeWidget extends StatelessWidget {
 
   void _onPopularTracks(BuildContext context) {
     pushSpiff(
+        ref: '/api/activity/tracks/popular/playlist',
         context,
         (ClientCubit client, {Duration? ttl}) =>
             client.popularTracks(ttl: Duration.zero));
+  }
+
+  void _onPlaylists(BuildContext context) {
+    push(context, builder: (_) => PlaylistsWidget());
   }
 
   void _onLink(BuildContext context) {